Custom prompt formatting via closure
